// src/enums/DiscordLimits.enum.ts
export enum DiscordLimits {
    GUILDS_PER_SHARD = 2500,
    CHANNELS_PER_GUILD = 500,
    ROLES_PER_GUILD = 250,
    PINS_PER_CHANNEL = 50,
    ACTIVE_THREADS_PER_GUILD = 1000,
    SELECT_MIN_ITEM_CHOSEN = 0,
    SELECT_MAX_ITEM_CHOSEN = 25,
    MODAL_TEXT_INPUT_MIN_VALUE = 0,
    MODAL_TEXT_INPUT_MAX_VALUE = 4000,
    EMBEDS_PER_MESSAGE = 10,
    FIELDS_PER_EMBED = 25,
    CHOICES_PER_AUTOCOMPLETE = 25,
    EMBED_COMBINED_LENGTH = 6000,
    EMBED_AUTHOR_LENGTH = 256,
    EMBED_TITLE_LENGTH = 256,
    EMBED_DESCRIPTION_LENGTH = 4096,
    EMBED_FIELD_NAME_LENGTH = 256,
    EMBED_FIELD_VALUE_LENGTH = 1024,
    EMBED_FOOTER_LENGTH = 2048,
}