// src/enums/TickEmoji.enum.ts
export enum TickEmoji {
    CHECK = '<:check:1244583873690599505>',
    CROSS = '<:cross:1244583872063213591>',
    PREV = '<:prev:1244583879361167370>',
    NEXT = '<:next:1244583870112727040>',
    UP = '<:up:1244583875087433738>',
    DOWN = '<:down:1244583877423398962>',
    BACK = '<:back:1218922818272624742>',
    REFRESH = '<:refresh:1247407727928410193>',
    BRUSH = '<:brush:1249621909361000520>',
    WRITE = '<:write:1249621914855673919>'
}