// src/modules/clan/ad/enums/CustomIds.enum.ts
export const CLAN_AD: string = 'j5+D3r-EWtSpF-AS6Fpc-:Mhe5s'
export const CLAN_AD_SEND: string = 'HcEAG8-AFw@6~-cLWkAv-mrP3c3'
export const CLAN_AD_NAME: string = '4dL4x)-eSZZBa-eMELvt-7)Kf8h'
export const CLAN_AD_NAME_MODAL: string = 'a**mta-T4HkFp-hyS5s4-V5dBHm'
export const CLAN_AD_NAME_MODAL_VALUE: string = 'x82+DK-ueFC3^-KzFKbA-sv2PaK'
export const CLAN_AD_DELETE: string = 'xrMgSH-eLC3kS-Zsv:C7-GJ$49S'
export const CLAN_AD_SELECT: string = 'xK9HSe-{j4haG-6MMEr4-f|wpFP'
export const CLAN_AD_CREATE: string = '2Wm5Cz-^3qjEP-RWzfz8-VF?ScT'
export const CLAN_AD_CREATE_MODAL: string = '4JBaQc-+9EQuK-.4xkYV-CwuQT7'
export const CLAN_AD_CREATE_MODAL_VALUE: string = 'FNChra-D+i79.-J6vrHh-wKpJ6E'
export const CLAN_AD_EDIT: string = 'Lte}Qu-uzABr8-CQn7N4-Kc7k?A'
export enum ClanAdEditOptions {
    Author = 'author',
    Body = 'body',
    Fields = 'fields',
    Images = 'images',
    Footer = 'footer',
}
export const CLAN_AD_EDIT_FIELDS_SELECT: string = '?LkBy|-fcLVaS-Avx57Z-AWE97T'
export const CLAN_AD_EDIT_FIELD_ADD: string = 'jSPAzF-3bU5v8-dHV3d]-vFik&S'
export const CLAN_AD_EDIT_FIELD_ADD_MODAL: string = 'yeLN6j-$CxUbD-{7XJcG-Erw39C'
export const CLAN_AD_EDIT_FIELD_DELETE: string = 'gR:aG2-5UuAAh-dH8qhL-3waV.Q'
export const CLAN_AD_EDIT_FIELD_MOVE_UP: string = 'Ph}Zye-uY2kBX-2b6}gG-x2qYkk'
export const CLAN_AD_EDIT_FIELD_MOVE_DOWN: string = 'pzhFsS->w7DX3-Ahmi(4-HDVQB2'
export enum ClanAdEditButtons {
    author = '7g5vDQ-73ahyT-*EzNyf-JeJ~Lr',
    body = 'efh)VD-gSrf4h-X>sDF4-hd86TC',
    fields = 'B!fAy7-dHMNYm-kC4X66-.bAXtM',
    images = 'n_Sewz-Uu6V3p-4FSAg7-FD=Kgd',
    footer = 'PH2jai-833AaG-PpacJh-=PcG,r',
}
export enum ClanAdEditAuthorModal {
    modalId = 'cGG4qw-ArXwJ4-sq}N]5-5yAAPn',
    name = 'name',
    url = 'url',
    iconUrl = 'iconUrl'
}
export enum ClanAdEditBodyModal {
    modalId = 'bV4hxY-sb98SY-m4#ukc-iySYf`',
    color = 'color',
    title = 'title',
    description = 'description',
    url = 'url',
}
export enum ClanAdEditFieldModal {
    modalId = '3]YikN-i=LmPK-i3SuPz-4n8YQx',
    name = 'name',
    value = 'value',
    inline = 'inline',
}
export enum ClanAdEditImagesModal {
    modalId = 'A[hDWr-V32cdg-TDV5cU-2,dyiv',
    image = 'image',
    thumbnail = 'thumbnail',
}
export enum ClanAdEditFooterModal {
    modalId = 'KDVbS3-nEsm3Y-FLHkjw-3bD$`4',
    text = 'text',
    iconUrl = 'iconUrl',
}